fix: refactor into separate files
