+++ nibbles/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'navigation/app_navigation.dart';
 
 void main() {
   runApp(const MainApp());
@@ -12,91 +13,6 @@ class MainApp extends StatelessWidget {
     return MaterialApp(
       theme: ThemeData(useMaterial3: true),
       home: const AppNavigation(),
-    );
-  }
-}
-
-class AppNavigation extends StatefulWidget {
-  const AppNavigation({super.key});
-
-  @override
-  State<AppNavigation> createState() => _AppNavigationState();
-}
-
-class _AppNavigationState extends State<AppNavigation> {
-  int currentPageIndex = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    final ThemeData theme = Theme.of(context);
-    return Scaffold(
-      bottomNavigationBar: NavigationBar(
-        onDestinationSelected: (int index) {
-          setState(() {
-            currentPageIndex = index;
-          });
-        },
-        indicatorColor: Colors.redAccent,
-        selectedIndex: currentPageIndex,
-        destinations: const <Widget>[
-          NavigationDestination(
-            selectedIcon: Icon(Icons.home),
-            icon: Icon(Icons.home_outlined),
-            label: 'Home',
-          ),
-          NavigationDestination(
-            selectedIcon: Icon(Icons.restaurant),
-            icon: Icon(Icons.restaurant_outlined),
-            label: 'Restaurants',
-          ),
-          NavigationDestination(
-            selectedIcon: Icon(Icons.menu_book),
-            icon: Icon(Icons.menu_book_outlined),
-            label: 'Recipes',
-          ),
-        ],
-      ),
-      body:
-          <Widget>[
-            /// Home page
-            Card(
-              shadowColor: Colors.transparent,
-              margin: const EdgeInsets.all(8.0),
-              child: SizedBox.expand(
-                child: Center(
-                  child: Text('Home page', style: theme.textTheme.titleLarge),
-                ),
-              ),
-            ),
-
-            /// Restaurants page
-            Card(
-              shadowColor: Colors.transparent,
-              margin: const EdgeInsets.all(8.0),
-              child: SizedBox.expand(
-                child: Center(
-                  child: Text(
-                    'Restaurants page',
-                    style: theme.textTheme.titleLarge,
-                  ),
-                ),
-              ),
-            ),
-
-            /// Recipes page
-            Card(
-              shadowColor: Colors.transparent,
-              margin: const EdgeInsets.all(8.0),
-              child: SizedBox.expand(
-                child: Center(
-                  child: Text(
-                    'Recipes page',
-                    style: theme.textTheme.titleLarge,
-                  ),
-                ),
-              ),
-            ),
-          ][currentPageIndex],
     );
   }
 }
