+++ nibbles/lib/navigation/app_navigation.dart
@@ -1,0 +1,53 @@
+import 'package:flutter/material.dart';
+import '../pages/home_page.dart';
+import '../pages/restaurants_page.dart';
+import '../pages/recipes_page.dart';
+
+class AppNavigation extends StatefulWidget {
+  const AppNavigation({super.key});
+
+  @override
+  State<AppNavigation> createState() => _AppNavigationState();
+}
+
+class _AppNavigationState extends State<AppNavigation> {
+  int currentPageIndex = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      bottomNavigationBar: NavigationBar(
+        onDestinationSelected: (int index) {
+          setState(() {
+            currentPageIndex = index;
+          });
+        },
+        indicatorColor: Colors.redAccent,
+        selectedIndex: currentPageIndex,
+        destinations: const <Widget>[
+          NavigationDestination(
+            selectedIcon: Icon(Icons.home),
+            icon: Icon(Icons.home_outlined),
+            label: 'Home',
+          ),
+          NavigationDestination(
+            selectedIcon: Icon(Icons.restaurant),
+            icon: Icon(Icons.restaurant_outlined),
+            label: 'Restaurants',
+          ),
+          NavigationDestination(
+            selectedIcon: Icon(Icons.menu_book),
+            icon: Icon(Icons.menu_book_outlined),
+            label: 'Recipes',
+          ),
+        ],
+      ),
+      body:
+          <Widget>[
+            const HomePage(),
+            const RestaurantsPage(),
+            const RecipesPage(),
+          ][currentPageIndex],
+    );
+  }
+}
