+++ nibbles/lib/pages/home_page.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    return Card(
+      shadowColor: Colors.transparent,
+      margin: const EdgeInsets.all(8.0),
+      child: SizedBox.expand(
+        child: Center(
+          child: Text('Home page', style: theme.textTheme.titleLarge),
+        ),
+      ),
+    );
+  }
+}
